webIntegration/package/microServiceV2/microServiceV2/pyproject.toml
[tool.poetry]
name = "microServiceV2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.4"
sklearn = "^0.0"
torch = "^1.11.0"
torchvision = "^0.12.0"
Pillow = "^9.1.1"
pillow-heif = "^0.2.5"
opencv-python = "^4.6.0"
pandas = "^1.4.2"
PyYAML = "^6.0"
tqdm = "^4.64.0"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
